refactor(docs): alert dx
