+++ apps/docs/content/components/alert/radius.raw.jsx
@@ -1,0 +1,15 @@
+import {Alert} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex items-center justify-center w-full">
+      <div className="flex flex-col w-full">
+        {["none", "sm", "md", "lg", "full"].map((radius) => (
+          <div key={radius} className="w-full flex items-center my-3">
+            <Alert radius={radius} title={`This is a ${radius} radius alert`} />
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
