+++ apps/docs/content/components/alert/usage.ts
@@ -1,15 +1,4 @@
-const App = `import {Alert} from "@nextui-org/react";
-
-export default function App() {
-  const title = "This is an alert";
-  const description = "Thanks for subscribing to our newsletter!";
-  
-  return (
-    <div className="flex items-center justify-center w-full">
-      <Alert title={title} description={description} />
-    </div>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
