+++ apps/docs/content/components/alert/without-icon.ts
@@ -1,21 +1,4 @@
-const App = `import {Alert} from "@nextui-org/react";
-
-export default function App() {
-  const title = "This is an alert";
-  const description = "Thanks for subscribing to our newsletter!";
-  
-  return (
-    <div className="flex items-center justify-center w-full">
-      <Alert 
-        title={title} 
-        description={description} 
-        hideIcon 
-        color="success" 
-        variant="faded" 
-      />
-    </div>
-  );
-}`;
+import App from "./without-icon.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
